fix: static 메서드 대신 인스턴스 메서드 또는 함수형 방식 사용, enumconverter utils에 따로 관리
